+++ src/apps/applib/lib/utility.c
@@ -18,10 +18,12 @@
  */
 
 #include "utility.h"
+#include <assert.h>
 #include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include "geoCoord.h"
+#include "h3Index.h"
 #include "h3api.h"
 
 void error(const char* msg) {
@@ -162,4 +164,36 @@ int readBoundary(FILE* f, GeoBoundary* b) {
     }
 
     return 0;
+}
+
+/**
+ * Call the callback for every index at the given resolution.
+ */
+void iterateAllIndexesAtRes(int res, void (*callback)(H3Index)) {
+    iterateAllIndexesAtResPartial(res, callback, NUM_BASE_CELLS);
+}
+
+/**
+ * Call the callback for every index at the given resolution in base
+ * cell 0 up to the given base cell number.
+ */
+void iterateAllIndexesAtResPartial(int res, void (*callback)(H3Index),
+                                   int baseCells) {
+    assert(baseCells <= NUM_BASE_CELLS);
+    for (int i = 0; i < baseCells; i++) {
+        H3Index bc;
+        setH3Index(&bc, 0, i, 0);
+        int childrenSz = H3_EXPORT(maxUncompactSize)(&bc, 1, res);
+        H3Index* children = calloc(childrenSz, sizeof(H3Index));
+        H3_EXPORT(uncompact)(&bc, 1, children, childrenSz, res);
+
+        for (int j = 0; j < childrenSz; j++) {
+            if (children[j] == 0) {
+                continue;
+            }
+
+            (*callback)(children[j]);
+        }
+        free(children);
+    }
 }