+++ src/h3lib/include/h3api.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2017 Uber Technologies, Inc.
+ * Copyright 2016-2018 Uber Technologies, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -128,6 +128,15 @@ struct LinkedGeoPolygon {
     LinkedGeoLoop *last;
     LinkedGeoPolygon *next;
 };
+
+/** @struct CoordIJK
+ * @brief IJK hexagon coordinates
+ */
+typedef struct {
+    int i;  ///< i component
+    int j;  ///< j component
+    int k;  ///< k component
+} CoordIJK;
 
 /** @defgroup geoToH3 geoToH3
  * Functions for geoToH3
@@ -439,6 +448,30 @@ void H3_EXPORT(getH3UnidirectionalEdgesFromHexagon)(H3Index origin,
  */
 /** @brief Returns the GeoBoundary containing the coordinates of the edge */
 void H3_EXPORT(getH3UnidirectionalEdgeBoundary)(H3Index edge, GeoBoundary *gb);
+/** @} */
+
+/** @defgroup ijkDistance ijkDistance
+ * Functions for ijkDistance
+ * @{
+ */
+/** @brief Returns the distance in cells between the two coordinates */
+int H3_EXPORT(ijkDistance)(const CoordIJK *a, const CoordIJK *b);
+/** @} */
+
+/** @defgroup h3ToIjk h3ToIjk
+ * Functions for h3ToIjk
+ * @{
+ */
+/** @brief Returns planar coordinates for an index */
+int H3_EXPORT(h3ToIjk)(H3Index origin, H3Index h3, CoordIJK *out);
+/** @} */
+
+/** @defgroup h3Distance h3Distance
+ * Functions for h3Distance
+ * @{
+ */
+/** @brief Returns grid distance between two indexes */
+int H3_EXPORT(h3Distance)(H3Index origin, H3Index h3);
 /** @} */
 
 #ifdef __cplusplus